requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.11.0
astunparse==1.6.3
asyncpg==0.30.0
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
elevenlabs==2.8.1
elevenlabs-mcp==0.5.1
fastapi==0.109.2
filelock==3.19.1
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.59.2
frozenlist==1.8.0
fsspec==2025.9.0
fuzzywuzzy==0.18.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
grpcio==1.74.0
gunicorn==23.0.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface_hub==1.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
keras==3.11.3
kiwisolver==1.4.9
Levenshtein==0.27.1
libclang==18.1.1
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mcp==1.12.3
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
multidict==6.7.0
namex==0.1.0
narwhals==2.5.0
networkx==3.5
numpy==2.2.4
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.2
pillow==11.3.0
propcache==0.4.1
protobuf==6.32.1
pyarrow==21.0.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.27.1
python-multipart==0.0.20
pytz==2025.2
pywin32==311
PyYAML==6.0.3
RapidFuzz==3.13.0
referencing==0.36.2
requests==2.32.4
rich==14.1.0
rpds-py==0.26.0
safetensors==0.7.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.1
sqlparse==0.5.4
sse-starlette==3.0.2
starlette==0.36.3
streamlit==1.49.1
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
timm==1.0.22
toml==0.10.2
torch==2.9.1
torchvision==0.24.1
tornado==6.5.2
tqdm==4.67.1
typer==0.16.0
typer-slim==0.20.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.27.1
watchdog==6.0.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
yarl==1.22.0
